+++ src/ext4_blockdev.c
@@ -115,7 +115,7 @@ int ext4_block_bind_bcache(struct ext4_blockdev *bdev, struct ext4_bcache *bc)
 	return EOK;
 }
 
-void ext4_block_set_lb_size(struct ext4_blockdev *bdev, uint64_t lb_bsize)
+void ext4_block_set_lb_size(struct ext4_blockdev *bdev, uint32_t lb_bsize)
 {
 	/*Logical block size has to be multiply of physical */
 	ext4_assert(!(lb_bsize % bdev->bdif->ph_bsize));
